Time: 197 ms (11.47%), Space: 86.6 MB (6.47%) - LeetHub

+++ 1803-average-waiting-time/1803-average-waiting-time.cpp
@@ -3,7 +3,7 @@ public:
     double averageWaitingTime(vector<vector<int>>& customers) {
 
         int availableAt = 0;
-        long waitingTime = 0;
+        double waitingTime = 0;
         for(auto customer: customers){
             int arrival = customer[0];
             int time = customer[1];
